+++ tests/testthat/test-bioconductor.R
@@ -185,3 +185,54 @@ test_that("auto-bioc install happens silently", {
   expect_true(renv_package_installed("BiocManager"))
 
 })
+
+test_that("standard bioc remotes are standardized appropriately", {
+
+  contents <- heredoc('
+    Package: BiocVersion
+    Version: 3.18.1
+    Title: Set the appropriate version of Bioconductor packages
+    Description: This package provides repository information for the appropriate version of Bioconductor.
+    Authors@R: c( person("Martin", "Morgan", email = "martin.morgan@roswellpark.org", role = "aut"), person("Marcel",
+                 "Ramos", email = "marcel.ramos@roswellpark.org", role = "ctb"), person("Bioconductor", "Package
+                 Maintainer", email = "maintainer@bioconductor.org", role = c("ctb", "cre")))
+    biocViews: Infrastructure
+    Depends: R (>= 4.3.0)
+    License: Artistic-2.0
+    Encoding: UTF-8
+    RoxygenNote: 6.0.1
+    git_url: https://git.bioconductor.org/packages/BiocVersion
+    git_branch: RELEASE_3_18
+    git_last_commit: 70680b8
+    git_last_commit_date: 2023-11-15
+    Repository: Bioconductor 3.18
+    Date/Publication: 2023-11-18
+    NeedsCompilation: no
+    Packaged: 2023-11-18 19:15:45 UTC; biocbuild
+    Author: Martin Morgan [aut], Marcel Ramos [ctb], Bioconductor Package Maintainer [ctb, cre]
+    Maintainer: Bioconductor Package Maintainer <maintainer@bioconductor.org>
+    Built: R 4.3.2; ; 2023-11-20 12:36:26 UTC; unix
+    RemoteType: standard
+    RemotePkgRef: BiocVersion
+    RemoteRef: BiocVersion
+    RemoteRepos: https://bioconductor.org/packages/3.18/bioc
+    RemotePkgPlatform: aarch64-apple-darwin20
+    RemoteSha: 3.18.1
+  ')
+
+  descfile <- tempfile("biocversion-")
+  writeLines(contents, con = descfile)
+
+  actual <- renv_snapshot_description(path = descfile)
+  expected <- list(
+    Package      = "BiocVersion",
+    Version      = "3.18.1",
+    Source       = "Bioconductor",
+    Repository   = "Bioconductor 3.18",
+    Requirements = "R",
+    Hash         = "2ecaed86684f5fae76ed5530f9d29c4a"
+  )
+
+  expect_identical(actual, expected)
+
+})